Migrate scroll-area from /docs/components/ to /components/
Port Horizontal Scrolling and Both Axes examples to the new reference
page, remove the old documentation page, and update all navigation
links and E2E tests.
