+++ site/ui/pages/components/scroll-area.tsx
@@ -5,7 +5,11 @@
  * Part of the #515 page redesign initiative.
  */
 
-import { ScrollAreaTagsDemo } from '@/components/scroll-area-demo'
+import {
+  ScrollAreaTagsDemo,
+  ScrollAreaHorizontalDemo,
+  ScrollAreaBothAxesDemo,
+} from '@/components/scroll-area-demo'
 import { ScrollAreaPlayground } from '@/components/scroll-area-playground'
 import {
   DocPage,
@@ -23,6 +27,9 @@ const tocItems: TocItem[] = [
   { id: 'preview', title: 'Preview' },
   { id: 'installation', title: 'Installation' },
   { id: 'usage', title: 'Usage' },
+  { id: 'examples', title: 'Examples' },
+  { id: 'horizontal', title: 'Horizontal Scrolling', branch: 'start' },
+  { id: 'both-axes', title: 'Both Axes', branch: 'end' },
   { id: 'api-reference', title: 'API Reference' },
 ]
 
@@ -46,6 +53,51 @@ function ScrollAreaDemo() {
             <Separator className="my-2" />
           </div>
         ))}
+      </div>
+    </ScrollArea>
+  )
+}`
+
+const horizontalCode = `"use client"
+
+import { ScrollArea } from "@/components/ui/scroll-area"
+
+function ScrollAreaHorizontal() {
+  return (
+    <ScrollArea class="w-96 whitespace-nowrap rounded-md border">
+      <div className="flex w-max space-x-4 p-4">
+        {works.map((work) => (
+          <figure className="shrink-0">
+            <div className="overflow-hidden rounded-md">
+              <div className="h-[150px] w-[200px] bg-muted" />
+            </div>
+            <figcaption className="pt-2 text-xs text-muted-foreground">
+              Photo by <span className="font-semibold text-foreground">{work.artist}</span>
+            </figcaption>
+          </figure>
+        ))}
+      </div>
+    </ScrollArea>
+  )
+}`
+
+const bothAxesCode = `"use client"
+
+import { ScrollArea } from "@/components/ui/scroll-area"
+
+function ScrollAreaBothAxes() {
+  return (
+    <ScrollArea class="h-64 w-80 rounded-md border">
+      <div className="p-4" style="width: 600px;">
+        <h4 className="mb-4 text-sm font-medium leading-none">Changelog</h4>
+        <div className="space-y-4">
+          {entries.map((entry) => (
+            <div className="whitespace-nowrap">
+              <div className="text-sm font-medium">{entry.title}</div>
+              <p className="text-sm text-muted-foreground">{entry.description}</p>
+            </div>
+          ))}
+        </div>
       </div>
     </ScrollArea>
   )
@@ -89,6 +141,19 @@ export function ScrollAreaRefPage() {
           <Example title="" code={usageCode}>
             <ScrollAreaTagsDemo />
           </Example>
+        </Section>
+
+        {/* Examples */}
+        <Section id="examples" title="Examples">
+          <div className="space-y-8">
+            <Example title="Horizontal Scrolling" code={horizontalCode}>
+              <ScrollAreaHorizontalDemo />
+            </Example>
+
+            <Example title="Both Axes" code={bothAxesCode}>
+              <ScrollAreaBothAxesDemo />
+            </Example>
+          </div>
         </Section>
 
         {/* API Reference */}
